+++ Inspector/PolygonInspector.cs
@@ -1,4 +1,4 @@
-//
+﻿//
 // Copyright (c) 2017 Geri Borbás http://www.twitter.com/_eppz
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
@@ -7,53 +7,17 @@
 //
 using UnityEngine;
 using System.Collections;
-using UnityEngine.Serialization;
 
 
-
-namespace EPPZ.Geometry.Components
+namespace EPPZ.Geometry.Inspector
 {
-
-
-	public class PolygonSource : MonoBehaviour
+	
+	
+	public class PolygonInspector : MonoBehaviour
 	{
 
 
-		public Transform[] pointTransforms;
-		public float offset = 0.0f;
-
-		public enum UpdateMode { Awake, Update, LateUpdate };
-		public UpdateMode update = UpdateMode.Awake;		
-
-		Polygon _polygon;
-		Polygon _offsetPolygon;		
-		public Polygon polygon { get { return (offset != 0.0f) ? _offsetPolygon : _polygon; } }
-
-
-		void Awake()
-		{
-			// Construct a polygon model from transforms (if not created by a root polygon already).
-			if (_polygon == null) _polygon = Polygon.PolygonWithSource(this);
-			if (offset != 0.0f) _offsetPolygon = _polygon.OffsetPolygon(offset);
-		}
-
-		void Update()
-		{
-			if (update == UpdateMode.Update)
-			{ UpdateModel(); }
-		}
-
-		void LateUpdate()
-		{
-			if (update == UpdateMode.LateUpdate)
-			{ UpdateModel(); }
-		}
-
-		void UpdateModel()
-		{
-			// Update polygon model with transforms, also update calculations.
-			_polygon.UpdatePointPositionsWithSource(this);
-			if (offset != 0.0f) _offsetPolygon = _polygon.OffsetPolygon(offset);
-		}
+		public Model.Polygon polygon;
+		public int currentEdgeIndex = 0;
 	}
 }