+++ Lines/GeometryLineRenderer.cs
@@ -15,6 +15,7 @@ namespace EPPZ.Geometry.Lines
 
 
 	using EPPZ.Lines;
+	using Model;
 
 
 	public class GeometryLineRenderer : DirectLineRenderer