+++ Source/Polygon.cs
@@ -1,4 +1,4 @@
-﻿//
+//
 // Copyright (c) 2017 Geri Borbás http://www.twitter.com/_eppz
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
@@ -7,28 +7,35 @@
 //
 using UnityEngine;
 using System.Collections;
+using UnityEngine.Serialization;
 
 
-namespace EPPZ.Geometry.Components
+
+namespace EPPZ.Geometry.Source
 {
 
 
-	public class SegmentSource : MonoBehaviour
+	public class Polygon : MonoBehaviour
 	{
 
 
-		public Transform[] pointTransforms;
+		[UnityEngine.Serialization.FormerlySerializedAs("pointTransforms")]
+		public Transform[] points;
+		public float offset = 0.0f;
 
 		public enum UpdateMode { Awake, Update, LateUpdate };
-		public UpdateMode update = UpdateMode.Awake;	
+		public UpdateMode update = UpdateMode.Awake;		
 
-		public Segment segment;
+		Model.Polygon _polygon;
+		Model.Polygon _offsetPolygon;		
+		public Model.Polygon polygon { get { return (offset != 0.0f) ? _offsetPolygon : _polygon; } }
 
 
 		void Awake()
 		{
-			// Construct a segment model from transforms.
-			segment = Segment.SegmentWithSource(this);
+			// Construct a polygon model from transforms (if not created by a root polygon already).
+			if (_polygon == null) _polygon = Model.Polygon.PolygonWithSource(this);
+			if (offset != 0.0f) _offsetPolygon = _polygon.OffsetPolygon(offset);
 		}
 
 		void Update()
@@ -45,8 +52,9 @@ namespace EPPZ.Geometry.Components
 
 		void UpdateModel()
 		{
-			// Update segment model with transforms, also update calculations.
-			segment.UpdateWithSource(this);
+			// Update polygon model with transforms, also update calculations.
+			_polygon.UpdatePointPositionsWithSource(this);
+			if (offset != 0.0f) _offsetPolygon = _polygon.OffsetPolygon(offset);
 		}
 	}
 }